Fully implemented TATP loader... Now also using log4j... I gave up on using java.util.logging...
git-svn-id: https://oltpbenchmark.googlecode.com/svn/trunk@57 8c9255d6-0b70-036a-fcb6-8a999947b34c

+++ src/com/oltpbenchmark/benchmarks/tatp/TATPUtil.java
@@ -57,8 +57,8 @@ public abstract class TATPUtil {
     
     public static final Random rand = new Random();
 
-    public static int isActive() {
-        return number(1,100) < number(86,100) ? 1 : 0;
+    public static byte isActive() {
+        return (byte)(number(1,100) < number(86,100) ? 1 : 0);
     }
 
     public static Long getSubscriberId(long subscriberSize) {
